add a flake which provides a derivation and nix-develop settings
This adds a Nix flake, which can be ignored by those not interested in Nix.

The flake provides two large components: A derivation (akin to a package of the
TypeyPipe program itself) and a nix-develop ecosystem.  The former allows for a
bit-for-bit consistent build of the Typeypipe program on any Nix system, and the
nix-develop settings allow for development on the project with consistent
settings.